feat: 리뷰 점수 계산 추가

+++ src/main/java/com/project/baedalsodae/review/entity/Review.java
@@ -7,10 +7,11 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import java.util.UUID;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.util.UUID;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -22,11 +23,11 @@ public class Review extends BaseAuditEntity {
 
     private UUID userId;
     private UUID orderId;
-    private int rating;
+    private double rating;
     private String content;
     private boolean isHidden;
 
-    private Review(UUID userId, UUID orderId, int rating, String content) {
+    private Review(UUID userId, UUID orderId, double rating, String content) {
         validate(rating, content);
         this.userId = userId;
         this.orderId = orderId;
@@ -35,7 +36,7 @@ public class Review extends BaseAuditEntity {
         this.isHidden = false;
     }
 
-    private void validate(int rating, String content) {
+    private void validate(double rating, String content) {
         if (rating < 1 || rating > 5) {
             throw new BusinessException(ErrorCode.RATING_OUT_OF_RANGE);
         }
@@ -44,13 +45,13 @@ public class Review extends BaseAuditEntity {
         }
     }
 
-    public void update(int rating, String content) {
+    public void update(double rating, String content) {
         validate(rating, content);
         this.rating = rating;
         this.content = content;
     }
 
-    public static Review create(UUID userId, UUID orderId, int rating, String content) {
+    public static Review create(UUID userId, UUID orderId, double rating, String content) {
         return new Review(userId, orderId, rating, content);
     }
 }
